Change ClearedOrderSummary.selection_id type to int
Previously ClearedOrderSummary.selection_id was declared to be a float, however the
API documentation says that it is a long:

https://api.developer.betfair.com/services/webapps/docs/x/SIA6#BettingTypeDefinitions-ClearedOrderSummary

This changes the declared data type in the model to int. int is used
over long as this maintains Python 3 compatability (in Python 2
int('<some number>') will return a long if the number is big enough
to require it).

+++ betfair/models.py
@@ -325,7 +325,7 @@ class ClearedOrderSummary(BetfairModel):
     event_type_id = Field(DataType(six.text_type))
     event_id = Field(DataType(six.text_type))
     market_id = Field(DataType(six.text_type))
-    selection_id = Field(DataType(float))
+    selection_id = Field(DataType(int))
     handicap = Field(DataType(float))
     bet_id = Field(DataType(six.text_type))
     placed_data = Field(datetime_type)
